+++ SolutionItems/Documents/Release Notes.md
@@ -6,6 +6,7 @@
 LocationItem object was created to hold the newly created resource pools, so it can navigate back to correct Id that's been retrieved from the server
 * Server-side properties were added to javascript objects, except Navigation Properties (due to breezejs error)
 * Element.IsMainElement instead of ResourcePool MainElement (back to previous setup)
+* resourcePool.js - selectedElement() instead of CurrentElement prop
 
 **0.37**